+++ packages/apis/nestjs/src/modules/goods/export/service.ts
@@ -6,19 +6,23 @@ import {
   type IGoodsExportRecordStatus,
 } from '@xiaoshop/schema'
 import { Repository } from 'typeorm'
-import { Injectable } from '@nestjs/common'
+import { Inject, Injectable } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 import { EventEmitter2 } from '@nestjs/event-emitter'
 import { FailedException } from '~/common/exception'
 import { GoodsExportEvent } from '@/goods/goods.events'
 import { GoodsExportRecord } from '@/goods/export/entity'
 import { GetGoodsExportRecordPagesRequest } from '@/goods/export/dto'
+import { StaffLogService } from '@/staff/log/service'
 
 @Injectable()
 export class GoodsExportRecordService {
   constructor(
     @InjectRepository(GoodsExportRecord)
     private readonly repository: Repository<GoodsExportRecord>,
+
+    @Inject(StaffLogService)
+    private readonly log: StaffLogService,
 
     private readonly eventEmitter: EventEmitter2,
   ) {}
@@ -74,6 +78,8 @@ export class GoodsExportRecordService {
         GoodsExportEvent.name,
         new GoodsExportEvent(res.id, conditions),
       )
+
+      await this.log.write('商品管理', `导出记录:「${res.id}」`)
     }
     catch (e) {
       throw new FailedException('创建导出记录', e.message)
@@ -100,6 +106,8 @@ export class GoodsExportRecordService {
         count,
         result: result || '',
       })
+
+      await this.log.write('商品管理', `导出记录「${id}」${status === 'success' ? '成功' : '失败'}`)
     }
     catch (e) {
       throw new FailedException('更新导出记录', e.message)
@@ -114,7 +122,12 @@ export class GoodsExportRecordService {
    */
   async delete(id: number) {
     try {
-      await this.repository.delete(id)
+      const record = await this.repository.findOne({ where: { id } })
+
+      if (record) {
+        await this.repository.delete(id)
+        await this.log.write('商品管理', `删除导出记录「${id}」`)
+      }
     }
     catch (e) {
       throw new FailedException('删除导出记录', e.message)
